requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
httpx==0.27.2
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
pandas==2.2.3
anthropic==0.40.0
python-multipart==0.0.12
aiofiles==24.1.0
tenacity==9.0.0
